requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.7.2
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
boto3==1.28.28
botocore==1.31.28
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==5.1.1
Django==3.2.20
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-elasticsearch-dsl==8.0
django-environ==0.10.0
django-filter==23.3
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-dark-theme==0.3
elastic-transport==8.10.0
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
executing==1.2.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
ipython==8.14.0
jedi==0.19.0
jmespath==1.0.1
Markdown==3.5.1
matplotlib-inline==0.1.6
mysqlclient==2.2.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
whitenoise==6.5.0
zipp==3.17.0
